Simplify calculations, consolidate constants
