+++ time/build.gradle
@@ -17,33 +17,6 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-buildscript {
-    repositories {
-        jcenter()
-
-        mavenCentral()
-        mavenLocal()
-
-        // Spine releases repository.
-        maven { url = repos.spine }
-
-        // Spine snapshots repository.
-        maven { url = repos.spineSnapshots }
-    }
-
-    //noinspection GroovyAssignabilityCheck
-    dependencies {
-        classpath deps.build.guava
-        classpath(deps.build.gradlePlugins.protobuf) {
-            // exclude an old Guava version
-            exclude group: 'com.google.guava'
-        }
-    }
-    
-    configurations.all({
-        resolutionStrategy.cacheChangingModulesFor(0, 'seconds')
-    })
-}
 
 configurations {
     // Avoid collisions of Java classes defined both in `protobuf-lite` and `protobuf-java`
